# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Codon Inference API
  description: |
    Public API for the Codon Encoder inference service.

    This API provides access to codon-level analysis using a trained neural encoder
    that maps 64 codons to a hierarchical embedding space. The model learns a
    hierarchical structure that preserves biological relationships.

    **Important**: Raw model weights and 16-dimensional embeddings are not exposed.
    Only derived metrics (projections, confidence scores, cluster assignments)
    are available through this API.

    ## Features

    - **Sequence Encoding**: Convert DNA sequences to codon trajectories with projections
    - **Cluster Analysis**: Group codons by amino acid similarity
    - **Depth Analysis**: Analyze hierarchical structure at each depth level
    - **Synonymous Variants**: Generate alternative DNA sequences for proteins
    - **Visualization Data**: Full 64-codon point cloud with edges

    ## Rate Limits

    - 100 requests/minute for anonymous users
    - 1000 requests/minute with API key

  version: 1.0.0
  contact:
    name: AI Whisperers
    url: https://ai-whisperers.org
    email: api@ai-whisperers.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://codon.ai-whisperers.org/api
    description: Production server
  - url: http://localhost:8000/api
    description: Local development

tags:
  - name: Visualization
    description: Endpoints for visualization data
  - name: Inference
    description: Sequence encoding and analysis
  - name: Lookup
    description: Individual codon/cluster/depth queries
  - name: Analytics
    description: Statistical analysis endpoints

paths:
  /visualization:
    get:
      tags: [Visualization]
      summary: Get complete visualization data
      description: Returns all 64 codon points, edges, configuration, and metadata.
      operationId: getVisualization
      responses:
        '200':
          description: Visualization payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisualizationResponse'

  /points:
    get:
      tags: [Visualization]
      summary: Get all codon points
      description: Returns all 64 codon points with inference results.
      operationId: getPoints
      responses:
        '200':
          description: List of codon points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CodonPoint'

  /edges:
    get:
      tags: [Visualization]
      summary: Get edge connections
      description: Returns fiber/edge connections between codons.
      operationId: getEdges
      parameters:
        - name: mode
          in: query
          description: Connection mode
          schema:
            type: string
            enum: [hierarchical, amino_acid, depth, none]
            default: hierarchical
      responses:
        '200':
          description: List of edges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Edge'

  /metadata:
    get:
      tags: [Visualization]
      summary: Get model metadata
      description: Returns model version, scores, and configuration (no weights).
      operationId: getMetadata
      responses:
        '200':
          description: Model metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelMetadata'

  /encode:
    post:
      tags: [Inference]
      summary: Encode DNA sequence
      description: |
        Encode a DNA sequence into a codon trajectory.

        Returns inference results for each codon including:
        - 3D projections for visualization
        - Cluster assignments (amino acid predictions)
        - Confidence scores
        - Hierarchical depth information

        Non-ATCG characters are automatically filtered.
      operationId: encode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncodeRequest'
            examples:
              simple:
                summary: Start codon
                value:
                  sequence: "ATG"
              short_sequence:
                summary: Short gene fragment
                value:
                  sequence: "ATGGCTCTGTGG"
      responses:
        '200':
          description: Encoded codon trajectory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EncodedCodon'
        '400':
          description: Invalid sequence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /synonymous_variants:
    post:
      tags: [Inference]
      summary: Generate synonymous variants
      description: |
        Generate alternative DNA sequences that encode the same protein.

        Returns variants using different codon selection strategies:
        - **random**: Random codon selection per amino acid
        - **high_depth**: Prefer inner codons (deeper in hierarchy)
        - **low_depth**: Prefer outer codons (shallower in hierarchy)

        Each variant includes the full encoded trajectory and statistics.
      operationId: getSynonymousVariants
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SynonymousVariantsRequest'
            examples:
              short_protein:
                summary: Short peptide
                value:
                  protein: "MDDII"
                  n_variants: 3
      responses:
        '200':
          description: Generated variants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SynonymousVariantsResponse'
        '400':
          description: Invalid protein sequence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /codon/{codon}:
    get:
      tags: [Lookup]
      summary: Get single codon info
      description: Returns inference results for a specific codon.
      operationId: getCodon
      parameters:
        - name: codon
          in: path
          required: true
          description: Three-letter codon (e.g., ATG)
          schema:
            type: string
            pattern: '^[ATCG]{3}$'
            example: ATG
      responses:
        '200':
          description: Codon information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleCodonResponse'
        '404':
          description: Invalid codon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cluster/{idx}:
    get:
      tags: [Lookup]
      summary: Get cluster codons
      description: Returns all codons belonging to an amino acid cluster.
      operationId: getCluster
      parameters:
        - name: idx
          in: path
          required: true
          description: Cluster index (0-20)
          schema:
            type: integer
            minimum: 0
            maximum: 20
      responses:
        '200':
          description: Cluster information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterResponse'
        '404':
          description: Invalid cluster index
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /depth/{level}:
    get:
      tags: [Lookup]
      summary: Get depth level codons
      description: Returns all codons at a specific hierarchical depth.
      operationId: getDepth
      parameters:
        - name: level
          in: path
          required: true
          description: Depth level (0-9)
          schema:
            type: integer
            minimum: 0
            maximum: 9
      responses:
        '200':
          description: Depth level information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepthLevelResponse'
        '404':
          description: Invalid depth level
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /angular_variance:
    get:
      tags: [Analytics]
      summary: Get angular variance statistics
      description: |
        Returns circular variance metrics by depth level.

        Measures the separation quality between and within amino acid groups
        using angular statistics on the Poincaré disk.

        - **intra_aa_variance**: Variance within amino acid groups
        - **inter_aa_variance**: Variance between amino acid centroids
        - **ratio**: inter/intra (higher = better separation)
      operationId: getAngularVariance
      responses:
        '200':
          description: Angular variance statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AngularVarianceResponse'

  /reference/actb:
    get:
      tags: [Analytics]
      summary: Get ACTB reference
      description: |
        Returns the first 60 amino acids of beta-actin (ACTB)
        with three synonymous variant encodings for comparison.
      operationId: getReferenceActb
      responses:
        '200':
          description: ACTB variants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SynonymousVariantsResponse'

components:
  schemas:
    CodonPoint:
      type: object
      description: Single codon with inference results
      required:
        - codon
        - amino_acid
        - position
        - depth
        - projection
        - embedding_norm
        - poincare_radius
        - cluster_idx
        - confidence
        - margin
        - color
      properties:
        codon:
          type: string
          description: Three-letter codon
          example: ATG
        amino_acid:
          type: string
          description: Single-letter amino acid code
          example: M
        position:
          type: integer
          description: Position in canonical 64-codon ordering
          minimum: 0
          maximum: 63
        depth:
          type: integer
          description: Hierarchical depth level
          minimum: 0
          maximum: 9
        projection:
          type: array
          description: 3D projection coordinates
          items:
            type: number
          minItems: 3
          maxItems: 3
        embedding_norm:
          type: number
          description: Euclidean norm of internal embedding (scalar)
          minimum: 0
        poincare_radius:
          type: number
          description: Hyperbolic radius in Poincaré disk
          minimum: 0
          maximum: 1
        cluster_idx:
          type: integer
          description: Predicted amino acid cluster
          minimum: 0
          maximum: 20
        confidence:
          type: number
          description: Cluster assignment confidence
          minimum: 0
          maximum: 1
        margin:
          type: number
          description: Decision margin to second-nearest cluster
        color:
          type: string
          description: Display color (hex)
          example: "#FF5733"

    Edge:
      type: object
      required: [source, target, weight]
      properties:
        source:
          type: integer
          minimum: 0
          maximum: 63
        target:
          type: integer
          minimum: 0
          maximum: 63
        weight:
          type: number
          minimum: 0
          maximum: 1

    ModelMetadata:
      type: object
      required:
        - version
        - structure_score
        - cluster_accuracy
        - embed_dim
      properties:
        version:
          type: string
          description: Model version identifier
        structure_score:
          type: number
          description: Structure quality (Spearman correlation)
        cluster_accuracy:
          type: number
          description: Cluster classification accuracy
          minimum: 0
          maximum: 1
        embed_dim:
          type: integer
          description: Embedding dimensionality (informational)
        num_clusters:
          type: integer
          default: 21
        projection_method:
          type: string
          default: PCA

    EncodeRequest:
      type: object
      required: [sequence]
      properties:
        sequence:
          type: string
          description: DNA sequence (ATCG)
          minLength: 3
          example: ATGGCTCTGTGG

    EncodedCodon:
      type: object
      required:
        - seq_position
        - codon
        - amino_acid
        - ref_idx
        - projection
        - depth
        - poincare_radius
        - embedding_norm
        - cluster_idx
        - confidence
        - margin
      properties:
        seq_position:
          type: integer
          description: Position in input sequence
          minimum: 0
        codon:
          type: string
        amino_acid:
          type: string
        ref_idx:
          type: integer
          minimum: 0
          maximum: 63
        projection:
          type: array
          items:
            type: number
          minItems: 3
          maxItems: 3
        depth:
          type: integer
          minimum: 0
          maximum: 9
        poincare_radius:
          type: number
        embedding_norm:
          type: number
        cluster_idx:
          type: integer
        confidence:
          type: number
        margin:
          type: number

    SynonymousVariantsRequest:
      type: object
      required: [protein]
      properties:
        protein:
          type: string
          description: Protein sequence (single-letter AA codes)
          minLength: 1
          example: MDDII
        n_variants:
          type: integer
          default: 3
          minimum: 1
          maximum: 10

    SynonymousVariantsResponse:
      type: object
      required: [protein, variants]
      properties:
        protein:
          type: string
        variants:
          type: array
          items:
            $ref: '#/components/schemas/SynonymousVariant'

    SynonymousVariant:
      type: object
      required: [strategy, dna, encoded, stats]
      properties:
        strategy:
          type: string
          enum: [random, high_depth, low_depth]
        dna:
          type: string
        encoded:
          type: array
          items:
            $ref: '#/components/schemas/EncodedCodon'
        stats:
          $ref: '#/components/schemas/VariantStats'

    VariantStats:
      type: object
      required: [mean_depth, depth_std, mean_confidence]
      properties:
        mean_depth:
          type: number
        depth_std:
          type: number
        mean_confidence:
          type: number

    VisualizationResponse:
      type: object
      required: [points, edges, config, metadata]
      properties:
        points:
          type: array
          items:
            $ref: '#/components/schemas/CodonPoint'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/Edge'
        config:
          $ref: '#/components/schemas/VisualizationConfig'
        metadata:
          $ref: '#/components/schemas/ModelMetadata'

    VisualizationConfig:
      type: object
      required: [depth_colors, amino_acid_colors, fiber_threshold]
      properties:
        depth_colors:
          type: array
          items:
            type: string
          minItems: 10
          maxItems: 10
        amino_acid_colors:
          type: object
          additionalProperties:
            type: string
        fiber_threshold:
          type: number

    SingleCodonResponse:
      type: object
      required:
        - codon
        - amino_acid
        - position
        - depth
        - projection
        - poincare_radius
        - cluster_idx
        - confidence
        - margin
      properties:
        codon:
          type: string
        amino_acid:
          type: string
        position:
          type: integer
        depth:
          type: integer
        projection:
          type: array
          items:
            type: number
          minItems: 3
          maxItems: 3
        poincare_radius:
          type: number
        cluster_idx:
          type: integer
        confidence:
          type: number
        margin:
          type: number

    ClusterResponse:
      type: object
      required: [cluster_idx, amino_acid, codons]
      properties:
        cluster_idx:
          type: integer
        amino_acid:
          type: string
        codons:
          type: array
          items:
            $ref: '#/components/schemas/SingleCodonResponse'

    DepthLevelResponse:
      type: object
      required: [depth, poincare_radius, codons]
      properties:
        depth:
          type: integer
        poincare_radius:
          type: number
        codons:
          type: array
          items:
            $ref: '#/components/schemas/SingleCodonResponse'

    AngularVarianceResponse:
      type: object
      required: [by_depth, summary]
      properties:
        by_depth:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/DepthStats'
        summary:
          type: object
          properties:
            mean_intra:
              type: number
            mean_inter:
              type: number
            overall_ratio:
              type: number

    DepthStats:
      type: object
      required: [intra_aa_variance, inter_aa_variance, ratio, amino_acids]
      properties:
        intra_aa_variance:
          type: number
        inter_aa_variance:
          type: number
        ratio:
          type: number
        amino_acids:
          type: array
          items:
            type: string

    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
        status_code:
          type: integer

  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for higher rate limits

security:
  - {}
  - apiKey: []